perf(cli): defer heavy imports to reduce startup time
