+++ libs/cli/deepagents_cli/non_interactive.py
@@ -601,6 +601,8 @@ async def run_non_interactive(
     exit_stack = contextlib.ExitStack()
 
     if sandbox_type != "none":
+        # Conditional: sandbox_factory transitively imports provider modules
+        # and SDKs — skip that cost for the common no-sandbox path.
         from deepagents_cli.integrations.sandbox_factory import (  # noqa: PLC0415
             create_sandbox,
         )
